fix: update to newer ec-gpu version
The API of ec-gpu changed. The new version makes naming things easier,
as it can directly be derived from the fields. Also helper functions
where introduced so that less boilerplate is needed for generating
a GPU kernel.

BREAKING CHANGE: deeper down dependencies changed

While this change itself is not really a breaking change, updating
to the newer `blstrs` version is. The reason is that it depends on
a pinned version of `blst`, which was updated. If anything within
your dependency tree is using a different version of `blstrs` you
will end up hard to resolve dependency issues.